+++ gcs/src/main/java/com/google/cloud/hadoop/fs/gcs/GoogleHadoopFileSystemConfiguration.java
@@ -30,7 +30,6 @@ import com.google.cloud.hadoop.gcsio.GoogleCloudStorageOptions;
 import com.google.cloud.hadoop.gcsio.GoogleCloudStorageReadOptions;
 import com.google.cloud.hadoop.gcsio.GoogleCloudStorageReadOptions.Fadvise;
 import com.google.cloud.hadoop.gcsio.PerformanceCachingGoogleCloudStorageOptions;
-import com.google.cloud.hadoop.gcsio.authorization.AuthorizationHandler;
 import com.google.cloud.hadoop.gcsio.cooplock.CooperativeLockingOptions;
 import com.google.cloud.hadoop.util.AsyncWriteChannelOptions;
 import com.google.cloud.hadoop.util.AsyncWriteChannelOptions.PipeType;
@@ -413,17 +412,6 @@ public class GoogleHadoopFileSystemConfiguration {
   public static final HadoopConfigurationProperty<RedactedString> GCS_ENCRYPTION_KEY_HASH =
       new HadoopConfigurationProperty<>("fs.gs.encryption.key.hash");
 
-  /** Configuration for authorization handler implementation class. */
-  public static final HadoopConfigurationProperty<Class<? extends AuthorizationHandler>>
-      GCS_AUTHORIZATION_HANDLER_IMPL =
-          new HadoopConfigurationProperty<>("fs.gs.authorization.handler.impl");
-
-  /** Configuration prefix for custom authorization handler properties. */
-  public static final HadoopConfigurationProperty<Map<String, String>>
-      GCS_AUTHORIZATION_HANDLER_PROPERTIES_PREFIX =
-          new HadoopConfigurationProperty<>(
-              "fs.gs.authorization.handler.properties.", ImmutableMap.of());
-
   // TODO(b/120887495): This @VisibleForTesting annotation was being ignored by prod code.
   // Please check that removing it is correct, and remove this comment along with it.
   // @VisibleForTesting
@@ -477,12 +465,7 @@ public class GoogleHadoopFileSystemConfiguration {
         .setEncryptionAlgorithm(GCS_ENCRYPTION_ALGORITHM.get(config, config::get))
         .setEncryptionKey(RedactedString.create(GCS_ENCRYPTION_KEY.getPassword(config)))
         .setEncryptionKeyHash(RedactedString.create(GCS_ENCRYPTION_KEY_HASH.getPassword(config)))
-        .setGrpcEnabled(GCS_GRPC_ENABLE.get(config, config::getBoolean))
-        .setAuthorizationHandlerImplClass(
-            GCS_AUTHORIZATION_HANDLER_IMPL.get(
-                config, (k, d) -> config.getClass(k, d, AuthorizationHandler.class)))
-        .setAuthorizationHandlerProperties(
-            GCS_AUTHORIZATION_HANDLER_PROPERTIES_PREFIX.getPropsWithPrefix(config));
+        .setGrpcEnabled(GCS_GRPC_ENABLE.get(config, config::getBoolean));
   }
 
   private static PerformanceCachingGoogleCloudStorageOptions getPerformanceCachingOptions(